% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_is_get_utils_deprecated.R
\docType{methods}
\name{XMLName,XiMpLe.node-method}
\alias{XMLName,XiMpLe.node-method}
\alias{XiMpLe-deprecated}
\alias{XMLName<-,XiMpLe.node-method}
\alias{XMLAttrs,XiMpLe.node-method}
\alias{XMLAttrs<-,XiMpLe.node-method}
\alias{XMLChildren,XiMpLe.node-method}
\alias{XMLChildren,XiMpLe.doc-method}
\alias{XMLChildren<-,XiMpLe.node-method}
\alias{XMLChildren<-,XiMpLe.doc-method}
\alias{XMLValue,XiMpLe.node-method}
\alias{XMLValue<-,XiMpLe.node-method}
\alias{XMLFile,XiMpLe.doc-method}
\alias{XMLFile<-,XiMpLe.doc-method}
\alias{XMLDecl,XiMpLe.doc-method}
\alias{XMLDecl<-,XiMpLe.doc-method}
\alias{XMLDTD,XiMpLe.doc-method}
\alias{XMLDTD<-,XiMpLe.doc-method}
\alias{XMLScan,XiMpLe.node-method}
\alias{XMLScan,XiMpLe.doc-method}
\alias{XMLScan<-,XiMpLe.node-method}
\alias{XMLScan<-,XiMpLe.doc-method}
\alias{XMLScanDeep,XiMpLe.node-method}
\alias{XMLScanDeep,XiMpLe.doc-method}
\alias{pasteXML,XiMpLe.node-method}
\alias{pasteXML,XiMpLe.doc-method}
\title{Deprecated functions and methods}
\usage{
\S4method{XMLName}{XiMpLe.node}(obj)

\S4method{XMLName}{XiMpLe.node}(obj) <- value

\S4method{XMLAttrs}{XiMpLe.node}(obj)

\S4method{XMLAttrs}{XiMpLe.node}(obj) <- value

\S4method{XMLChildren}{XiMpLe.node}(obj)

\S4method{XMLChildren}{XiMpLe.doc}(obj)

\S4method{XMLChildren}{XiMpLe.node}(obj) <- value

\S4method{XMLChildren}{XiMpLe.doc}(obj) <- value

\S4method{XMLValue}{XiMpLe.node}(obj)

\S4method{XMLValue}{XiMpLe.node}(obj) <- value

\S4method{XMLFile}{XiMpLe.doc}(obj)

\S4method{XMLFile}{XiMpLe.doc}(obj) <- value

\S4method{XMLDecl}{XiMpLe.doc}(obj)

\S4method{XMLDecl}{XiMpLe.doc}(obj) <- value

\S4method{XMLDTD}{XiMpLe.doc}(obj)

\S4method{XMLDTD}{XiMpLe.doc}(obj) <- value

\S4method{XMLScan}{XiMpLe.node}(obj, name, as.list = FALSE)

\S4method{XMLScan}{XiMpLe.doc}(obj, name, as.list = FALSE)

\S4method{XMLScan}{XiMpLe.node}(obj, name) <- value

\S4method{XMLScan}{XiMpLe.doc}(obj, name) <- value

\S4method{XMLScanDeep}{XiMpLe.node}(obj, find = NULL, search = "attributes")

\S4method{XMLScanDeep}{XiMpLe.doc}(obj, find = NULL, search = "attributes")

\S4method{pasteXML}{XiMpLe.node}(obj, ...)

\S4method{pasteXML}{XiMpLe.doc}(obj, ...)
}
\arguments{
\item{obj}{An object of deprecated classes \code{XiMpLe.node} or \code{XiMpLe.doc}.}

\item{value}{No longer used.}

\item{name}{No longer used.}

\item{as.list}{No longer used.}

\item{find}{No longer used.}

\item{search}{No longer used.}

\item{...}{No longer used.}
}
\description{
These functions were used in earlier versions of the package but are now either
replaced or removed.
}
\details{
For methods this can also mean that the object class is deprecated and you are asked
to update old objects via \code{\link[XiMpLe:as_XiMpLe_node]{as_XiMpLe_node}} or
\code{\link[XiMpLe:as_XiMpLe_doc]{as_XiMpLe_doc}}.
}
