% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTDNode.R
\name{DTDNode}
\alias{DTDNode}
\title{Constructor function for XiMpLe.DTD objects}
\usage{
DTDNode(name, element, publicID = NULL, systemID = NULL, decl = NULL,
  local = NULL)
}
\arguments{
\item{name}{A character string, name of this element (i.e.,
      the XML tag identifier). Must be one of "!DOCTYPE", "!ENTITY", "!ELEMENT",
"!ATTLIST", or "!NOTATION".}

\item{element}{A character string, name of the element that is being defined (e.g.,
      its attributes).
If \code{name="!ELEMENT"} or \code{name="!ATTLIST"}, this represents the tag name,
      if \code{name="!ENTITY"} the entity name.
If \code{name="!DOCTYPE"}, it names the doctype,
      if \code{name="!NOTATION"} the declared notation name.}

\item{publicID}{A character string defining the public identifier of a doctype, e.g.,
      "-//W3C//DTD HTML 4.01 Transitional//EN".
Only valid if \code{name="!DOCTYPE"} or \code{name="!NOTATION"}.}

\item{systemID}{A character string defining the system identifier of a doctype,
      usually a URI like
"http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd". Only valid if \code{name="!DOCTYPE"} or \code{name="!NOTATION"}.}

\item{decl}{A named list of character vectors or XML comments.
If \code{name="!ATTLIST"},
      each vector must have one to three entries: The vector name is the attribute name,
the first value its type description,
      the optional second and third values either one of \code{"#REQUIRED"}, \code{"#IMPLIED"} or \code{"#FIXED"},
and/or a default value.
If \code{name="!ELEMENT"} or \code{name="!ENTITY"},
      each vector must have one entry describing valid child nodes or the entity, respectively.
Invalid if \code{name="!NOTATION"} or \code{name="!DOCTYPE"}.}

\item{local}{A list of objects of class XiMpLe.DTD or XML comments,
      only valid if \code{name="!DOCTYPE"}. These
elements will form the local subset of document type definitions.}
}
\value{
An object of class \code{\link[XiMpLe:XiMpLe.DTD-class]{XiMpLe.DTD}}.
}
\description{
Can be used to create document type declaration nodes.
}
\seealso{
\code{\link[XiMpLe:XMLTree]{XMLTree}}
}

