\name{show}
\alias{show}
\alias{show,-methods}
\alias{show,XiMpLe.doc-method}
\alias{show,XiMpLe.node-method}
\title{Show method for objects of class XiMpLe.doc}
\arguments{
  \item{object}{An object of class \code{XiMpLe.doc}}

  \item{object}{An object of class \code{XiMpLe.node}}
}
\description{
  Show method for S4 objects of class
  \code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}}

  Show method for S4 objects of class
  \code{\link[XiMpLe:XiMpLe.node-class]{XiMpLe.node}}
}
\seealso{
  \code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}}

  \code{\link[XiMpLe:XiMpLe.node-class]{XiMpLe.node}}
}
\keyword{methods}

