% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_method_04_validXML.R
\docType{methods}
\name{validXML}
\alias{validXML}
\alias{validXML,-methods}
\alias{validXML,XiMpLe.XML-method}
\alias{validXML,XiMpLe.doc-method}
\alias{validXML,XiMpLe.node-method}
\title{Validate S4 objects of XiMpLe XML classes}
\usage{
validXML(obj, validity, parent = NULL, children = TRUE, attributes = TRUE,
  warn = FALSE, section = parent)

\S4method{validXML}{XiMpLe.XML}(obj, validity, parent = NULL,
  children = TRUE, attributes = TRUE, warn = FALSE, section = parent)
}
\arguments{
\item{obj}{An object of class \code{XiMpLe.doc} or \code{XiMpLe.node}. If \code{parent=NULL}, this object
will be checked for validity, including its child nodes. If \code{parent} is either a character string
or another XiMpLe node, it will be checked whether \code{obj} is a valid child node of \code{parent}.}

\item{validity}{A list with validity information.}

\item{parent}{Either a character string (name of the parent node) or a XiMpLe node, whose name will be used
as name of the parent node.}

\item{warn}{Logical, whether invalid objects should cause a warning or stop with an error.}

\item{section}{Either a character string (name of the section) or a XiMpLe node, whose name will be used
as name of the XML section this check refers to. This is only relevant for warnings and error messages,
in case you want to use something different than the actual parent node name.}
}
\value{
Returns \code{TRUE} if tests pass, and depending on the setting of \code{warn} either \code{FALSE} or
   an error if a test fails.
}
\description{
Check whether objects of class \code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}}
or \code{\link[XiMpLe:XiMpLe.node-class]{XiMpLe.node}} have valid child nodes.
}
\details{
XiMpLe can't handle DOM specifications yet, but this method can be used to construct
validation schemes.
}
\seealso{
\code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}}
   \code{\link[XiMpLe:XiMpLe.node-class]{XiMpLe.node}}
}
\keyword{methods}

