% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XiMpLe-package.R
\docType{package}
\name{XiMpLe-package}
\alias{XiMpLe}
\alias{XiMpLe-package}
\title{\packageTitle{XiMpLe}}
\description{
\packageDescription{XiMpLe}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab XiMpLe\cr
Type: \tab Package\cr
Version: \tab 0.11-1\cr
Date: \tab 2022-11-17\cr
Depends: \tab R (>= 2.9.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://reaktanz.de/?c=hacking&s=XiMpLe\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://reaktanz.de/?c=hacking&s=XiMpLe}
  \item Report bugs at \url{https://github.com/rkward-community/XiMpLe/issues}
}

}
\author{
\packageAuthor{XiMpLe}

Maintainer: \packageMaintainer{XiMpLe}
}
