% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_04_XiMpLe.DTD.R
\docType{class}
\name{XiMpLe.DTD,-class}
\alias{XiMpLe.DTD,-class}
\alias{XiMpLe.DTD-class}
\title{Class XiMpLe.DTD}
\description{
This class is used to create XML document type defintions to be used in the \code{dtd} slot in
\code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}} objects,
      e.g. by calling \code{\link[XiMpLe:XMLTree]{XMLTree}}.
}
\details{
The name of a node defines how slots are used and how nodes can be nested. You should use the function
\code{\link[XiMpLe:DTDNode]{DTDNode}} to create elements of this class.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character string, name of this element (i.e.,
      the XML tag identifier). Must be one of "!DOCTYPE", "!ENTITY", "!ELEMENT",
"!ATTLIST", or "!NOTATION".}

\item{\code{element}}{A character string, name of the element that is being defined (e.g.,
      its attributes).
If \code{name="!ELEMENT"} or \code{name="!ATTLIST"}, this represents the tag name,
      if \code{name="!ENTITY"} the entity name.
If \code{name="!DOCTYPE"}, it names the doctype,
      if \code{name="!NOTATION"} the declared notation name.}

\item{\code{publicID}}{A character string defining the public identifier of a doctype,
      e.g., "-//W3C//DTD HTML 4.01 Transitional//EN".
Only valid if \code{name="!DOCTYPE"} or \code{name="!NOTATION"}.}

\item{\code{systemID}}{A character string defining the system identifier of a doctype,
      usually a URI like
"http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd". Only valid if \code{name="!DOCTYPE"} or \code{name="!NOTATION"}.}

\item{\code{decl}}{A named list of character vectors or XML comments.
If \code{name="!ATTLIST"},
      each vector must have one to three entries: The vector name is the attribute name,
the first value its type description,
      the optional second and third values either one of \code{"#REQUIRED"}, \code{"#IMPLIED"} or \code{"#FIXED"},
and/or a default value.
If \code{name="!ELEMENT"} or \code{name="!ENTITY"},
      each vector must have one entry describing valid child nodes or the entity, respectively.
Invalid if \code{name="!NOTATION"} or \code{name="!DOCTYPE"}.}

\item{\code{local}}{A list of objects of class XiMpLe.DTD or XML comments,
      only valid if \code{name="!DOCTYPE"}. This
slot will form the local subset of document type definitions.}
}}
\examples{
\dontrun{
XMLDTD("!ELEMENT", element="html", decl=list("(head, body)"))
XMLDTD("!ATTLIST", element="img", decl=list(
    src=c("CDATA", "#REQUIRED"),
    id=c("ID", "#IMPLIED"),
    sort=c("CDATA", "#FIXED", "\\"true\\"")
    print=c("(yes | no)", "yes")
  )
)
}
}
\keyword{classes}

