% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_03_XiMpLe.validity.R, R/XMLValidity.R, R/zzz_is_get_utils.R
\docType{class}
\name{XiMpLe.validity,-class}
\alias{XMLValidity}
\alias{XiMpLe.validity,-class}
\alias{XiMpLe.validity-class}
\alias{is.XiMpLe.validity}
\title{Class XiMpLe.validity}
\usage{
XMLValidity(children = NULL, attrs = NULL, allChildren = NULL,
  allAttrs = NULL)

is.XiMpLe.validity(x)
}
\arguments{
\item{children}{Named list of character vectors, where the element name defines the parent node
name and each character string a valid child node name.}

\item{attrs}{Named list of character vectors, where the element name defines the parent node
name and each character string a valid attribute name.}

\item{allChildren}{Character vector, names of globally valid child nodes for all nodes, if any.}

\item{allAttrs}{Character vector, names of globally valid attributes for all nodes, if any.}

\item{x}{An arbitrary \code{R} object.}
}
\value{
An object of class \code{\link[XiMpLe:XiMpLe.validity-class]{XiMpLe.validity}}
}
\description{
Used for objects that describe valid child nodes and attributes of XiMpLe.nodes.

Create validity definitions for XiMpLe nodes, to be used by
\code{\link[XiMpLe:validXML]{validXML}}.
}
\section{Slots}{

\describe{
\item{\code{children}}{Named list of character vectors, where the element name defines the parent node
name and each character string a valid child node name.}

\item{\code{attrs}}{Named list of character vectors, where the element name defines the parent node
name and each character string a valid attribute name.}

\item{\code{allChildren}}{Character vector, names of globally valid child nodes for all nodes, if any.}

\item{\code{allAttrs}}{Character vector, names of globally valid attributes for all nodes, if any.}
}}
\seealso{
\code{\link[XiMpLe:validXML]{validXML}}
}
\keyword{classes}

