% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_02_XiMpLe.doc.R, R/zzz_is_get_utils.R
\docType{class}
\name{XiMpLe_doc,-class}
\alias{XiMpLe_doc,-class}
\alias{XiMpLe_doc_old}
\alias{XiMpLe.doc-class}
\alias{XiMpLe.doc,-class}
\alias{XiMpLe_doc-class}
\alias{XiMpLe_doc}
\alias{is.XiMpLe.doc}
\alias{as_XiMpLe_doc}
\alias{as_XiMpLe_doc,XiMpLe.doc-method}
\alias{as_XiMpLe_doc,-methods}
\title{Classes XiMpLe_doc and XiMpLe.doc (old)}
\usage{
is.XiMpLe.doc(x)

as_XiMpLe_doc(obj, ...)

\S4method{as_XiMpLe_doc}{XiMpLe.doc}(obj, extra = list(), version = 2)
}
\arguments{
\item{x}{An arbitrary \code{R} object.}
}
\description{
This class is used for objects that are returned by \code{\link[XiMpLe:parseXMLTree]{parseXMLTree}}.
}
\details{
Class \code{XiMpLe.doc} is the older one,
      \code{XiMpLe_doc} was introduced with \code{XiMpLe} 0.11-1. It has two new slots
(\code{extra} and \code{version}) that would have made it impossible to load old objects without issues. \code{XiMpLe_doc}
inherits from \code{XiMpLe.doc}. You can convert old objects into valid new ones using the \code{as_XiMpLe_doc} method and
are also advised to do so,
      as the \code{XiMpLe.doc} class might become deprecated in future releases.

A contructor function \code{XiMpLe_doc(...)} is available to be used instead of \code{new("XiMpLe_doc",
      ...)}.

There's also \code{XiMpLe_doc_old(...)} to be used instead of \code{new("XiMpLe.doc",
      ...)}, but you should not use that any longer.
}
\section{Slots}{

\describe{
\item{\code{file}}{Character string, Name of the file.}

\item{\code{xml}}{Either a named list of character values (attributes for the XML declaration of the file),
or a list of XiMpLe.nodes with tags whose names must start with a "?".}

\item{\code{dtd}}{A named list, attributes for the doctype definition of the file.}

\item{\code{children}}{A list of objects of class \code{\link[XiMpLe:XiMpLe_node]{XiMpLe_node}} (\code{XiMpLe_doc} only),
      or
\code{\link[XiMpLe:XiMpLe.node]{XiMpLe.node}} (old, \code{XiMpLe.doc} only),
      representing the DOM structure of the XML document.}

\item{\code{extra}}{A named list that can be used to store additional information on a document (\code{XiMpLe_doc} only).}

\item{\code{version}}{A numeric integer,
      currently defaults to 2 (\code{XiMpLe_doc} only). This is a version number that can be used in the future
in combination with the added \code{extra} slot. Should that get sime supported values that are interpreted by package methods,
the version number will be increased and the differences documented here. You shouldn't set it manually.}
}}

\keyword{classes}
