% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_method_05_XMLgenerators.R
\docType{methods}
\name{XMLgenerators}
\alias{XMLgenerators}
\alias{XMLgenerators,-methods}
\alias{XMLgenerators,XiMpLe.validity-method}
\title{Generate XML generator functions from XiMpLe.valisity object}
\usage{
XMLgenerators(validity, prefix = "XML", checkValidity = TRUE,
  indent.by = "\\t", roxygenDocs = FALSE)

\S4method{XMLgenerators}{XiMpLe.validity}(validity, prefix = "XML",
  checkValidity = TRUE, indent.by = "\\t", roxygenDocs = FALSE)
}
\arguments{
\item{validity}{An dobject of class \code{XiMpLe.validity}.}

\item{prefix}{A character string to be used as a prefix for the resulting function names.}

\item{checkValidity}{Logical, whether all functions should include a check for valid XML.}

\item{indent.by}{A charachter string defining how indentation should be done.}

\item{roxygenDocs}{Logical,
      whether a skeleton for roxygen2-ish documentation should be added.}
}
\value{
A named vector of character strings.
}
\description{
Takes an object of class \code{\link[XiMpLe:XiMpLe.validity-class]{XiMpLe.validity}}
and turns it into a character vector of generator functions for each XML node that
was defined.
}
\details{
The resulting code follows these rules:
\itemize{
   \item{Each child node gets its own argument, except if there is only one valid child
     node. It will use the dots element instead then.}
   \item{Each attribute will also get its own argument.}
   \item{If \code{CheckValidity=TRUE}, one extra argument \code{validity will be added}}
   \item{All arguments are set to \code{NULL} by default}
}
}
\examples{
HTMLish <- XMLValidity(
   children=list(
     body=c("a", "p", "ol", "ul", "strong"),
     head=c("title"),
     html=c("head", "body"),
     li=c("a", "br", "strong"),
     ol=c("li"),
     p=c("a", "br", "ol", "ul", "strong"),
     ul=c("li")
   ),
   attrs=list(
     a=c("href", "name"),
     p=c("align")
   ),
   allChildren=c("!--"),
   allAttrs=c("id", "class"),
   empty=c("br")
)
XMLgenerators(HTMLish)
}
\seealso{
\code{\link[XiMpLe:XMLValidity]{XMLValidity}} and
   \code{\link[XiMpLe:XiMpLe.validity-class]{XiMpLe.validity}}
}
\keyword{methods}

