\docType{methods}
\name{getXMLName}
\alias{getXMLAttrs}
\alias{getXMLAttrs,-methods}
\alias{getXMLAttrs,XiMpLe.node-method}
\alias{getXMLChildren}
\alias{getXMLChildren,-methods}
\alias{getXMLChildren,XiMpLe.doc-method}
\alias{getXMLChildren,XiMpLe.node-method}
\alias{getXMLDecl}
\alias{getXMLDecl,-methods}
\alias{getXMLDecl,XiMpLe.doc-method}
\alias{getXMLDTD}
\alias{getXMLDTD,-methods}
\alias{getXMLDTD,XiMpLe.doc-method}
\alias{getXMLFile}
\alias{getXMLFile,-methods}
\alias{getXMLFile,XiMpLe.doc-method}
\alias{getXMLName}
\alias{getXMLName,-methods}
\alias{getXMLName,XiMpLe.node-method}
\alias{getXMLValue}
\alias{getXMLValue,-methods}
\alias{getXMLValue,XiMpLe.node-method}
\title{Getter methods for S4 objects of XiMpLe XML classes}
\arguments{
  \item{obj}{An object of class \code{XiMpLe.node} or
  \code{XiMpLe.doc}}
}
\description{
  Used to get certain slots from objects of class
  \code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}} and
  \code{\link[XiMpLe:XiMpLe.node-class]{XiMpLe.node}}.
}
\details{
  These are convenience methods to get slots from XML
  objects without using the \code{@} operator.
}
\seealso{
  \code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}}
  \code{\link[XiMpLe:XiMpLe.node-class]{XiMpLe.node}}
}
\keyword{methods}

