\name{show}
\alias{show}
\alias{show,-methods}
\alias{show,XiMpLe.doc-method}
\alias{show,XiMpLe.node-method}
\alias{show,XiMpLe.XML-method}
\title{Show method for S4 objects of XiMpLe XML classes}
\arguments{
  \item{object}{An object of class \code{XiMpLe.doc} or
  \code{XiMpLe.node}}
}
\description{
  Used to display objects of class
  \code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}} and
  \code{\link[XiMpLe:XiMpLe.node-class]{XiMpLe.node}}
}
\seealso{
  \code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}}
  \code{\link[XiMpLe:XiMpLe.node-class]{XiMpLe.node}}
}
\keyword{methods}

