% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01_method_03_show.R
\docType{methods}
\name{show,XiMpLe.XML-method}
\alias{show,-methods}
\alias{show,XiMpLe.XML-method}
\alias{show,XiMpLe.doc-method}
\alias{show,XiMpLe.node-method}
\title{Show method for S4 objects of XiMpLe XML classes}
\usage{
\S4method{show}{XiMpLe.XML}(object)
}
\arguments{
\item{object}{An object of class \code{XiMpLe.doc} or \code{XiMpLe.node}}
}
\description{
Used to display objects of class \code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}}
and \code{\link[XiMpLe:XiMpLe.node-class]{XiMpLe.node}}
}
\seealso{
\code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}}
   \code{\link[XiMpLe:XiMpLe.node-class]{XiMpLe.node}}
}
\keyword{methods}

