% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_method_03_show.R
\docType{methods}
\name{show,XiMpLe.XML-method}
\alias{show,XiMpLe.XML-method}
\alias{show,-methods}
\alias{show,XiMpLe_doc-method}
\alias{show,XiMpLe_node-method}
\alias{show,XiMpLe.doc-method}
\alias{show,XiMpLe.node-method}
\title{Show method for S4 objects of XiMpLe XML classes}
\usage{
\S4method{show}{XiMpLe.XML}(object)
}
\arguments{
\item{object}{An object of class \code{XiMpLe_doc} or \code{XiMpLe_node}}
}
\description{
Used to display objects of class \code{\link[XiMpLe:XiMpLe_doc-class]{XiMpLe_doc}}
and \code{\link[XiMpLe:XiMpLe_node-class]{XiMpLe_node}}
}
\seealso{
\code{\link[XiMpLe:XiMpLe_doc-class]{XiMpLe_doc}}
   \code{\link[XiMpLe:XiMpLe_node-class]{XiMpLe_node}}
}
\keyword{methods}
