\docType{package}
\name{XiMpLe-package}
\alias{XiMpLe}
\alias{XiMpLe-package}
\title{The XiMpLe Package}
\description{
  A simple XML tree parser and generator.
}
\details{
  \tabular{ll}{ Package: \tab XiMpLe\cr Type: \tab
  Package\cr Version: \tab 0.03-18\cr Date: \tab
  2013-02-08\cr Depends: \tab R (>= 2.9.0),methods\cr
  Enhances: \tab rkward\cr Encoding: \tab UTF-8\cr License:
  \tab GPL (>= 3)\cr LazyLoad: \tab yes\cr URL: \tab
  http://reaktanz.de/?c=hacking&s=XiMpLe\cr }

  This package provides a simple XML tree parser/generator.
  It includes functions to read XML files into R objects,
  get information out of and into nodes, and write R
  objects back to XML code. It's not as powerful as the XML
  package and doesn't aim to be, but for simple XML
  handling it could be useful. It was originally programmed
  for RKWard.
}
\author{
  Meik Michalke \email{meik.michalke@hhu.de}
}
\keyword{package}

