% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XiMpLe-package.R
\docType{package}
\name{XiMpLe-package}
\alias{XiMpLe-package}
\title{The XiMpLe Package}
\description{
A Simple XML Tree Parser and Generator.
}
\details{
\tabular{ll}{
Package: \tab XiMpLe\cr
Type: \tab Package\cr
Version: \tab 0.03-24\cr
Date: \tab 2015-12-07\cr
Depends: \tab R (>= 2.9.0),methods\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://reaktanz.de/?c=hacking&s=XiMpLe\cr
}

Provides a simple XML tree parser/generator. It includes functions to read XML files into R objects,
get information out of and into nodes, and write R objects back to XML code.
It's not as powerful as the 'XML' package and doesn't aim to be, but for simple XML handling
it could be useful. It was originally developed for the R GUI and IDE RKWard (https://rkward.kde.org), to make plugin
development easier.
}
\author{
Meik Michalke
}
\keyword{package}

