\name{XiMpLe.doc-class}
\title{S4 class XiMpLe.doc}
\description{Class XiMpLe.doc}
\details{This class is used for objects that are returned by \code{\link[XiMpLe:parseXMLTree]{parseXMLTree}}.}
\alias{XiMpLe.doc-class}
\alias{XiMpLe.doc,-class}
\keyword{classes}
\section{Slots}{\describe{\item{\code{file}:}{(\code{\link{character}}) Name of the file.}\item{\code{xml}:}{(\code{\link{list}}) XML declaration of the file.}\item{\code{dtd}:}{(\code{\link{list}}) Doctype definition of the file.}\item{\code{children}:}{(\code{\link{list}}) A list of objects of class XiMpLe.node,
      representing the DOM structure of the XML document.}}}

