% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_02_XiMpLe.doc.R, R/zzz_is_get_utils.R
\docType{class}
\name{XiMpLe.doc,-class}
\alias{XiMpLe.doc,-class}
\alias{XiMpLe.doc-class}
\alias{is.XiMpLe.doc}
\title{Class XiMpLe.doc}
\usage{
is.XiMpLe.doc(x)
}
\arguments{
\item{x}{An arbitrary \code{R} object.}
}
\description{
This class is used for objects that are returned by \code{\link[XiMpLe:parseXMLTree]{parseXMLTree}}.
}
\section{Slots}{

\describe{
\item{\code{file}}{Character string, Name of the file.}

\item{\code{xml}}{Either a named list of character values (attributes for the XML declaration of the file),
or a list of XiMpLe.nodes with tags whose names must start with a "?".}

\item{\code{dtd}}{A named list, attributes for the doctype definition of the file.}

\item{\code{children}}{A list of objects of class XiMpLe.node,
      representing the DOM structure of the XML document.}
}}

\keyword{classes}
