% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_03_XiMpLe.validity.R, R/zzz_is_get_utils.R
\docType{class}
\name{XiMpLe.validity,-class}
\alias{XiMpLe.validity,-class}
\alias{XiMpLe.validity-class}
\alias{is.XiMpLe.validity}
\title{Class XiMpLe.validity}
\usage{
is.XiMpLe.validity(x)
}
\arguments{
\item{x}{An arbitrary \code{R} object.}
}
\description{
Used for objects that describe valid child nodes and attributes of XiMpLe.nodes.
}
\details{
You should use \code{\link[XiMpLe:XMLValidity]{XMLValidity}} to create objects of this class.
}
\section{Slots}{

\describe{
\item{\code{children}}{Named list of character vectors, where the element name defines the parent node
name and each character string a valid child node name.}

\item{\code{attrs}}{Named list of character vectors, where the element name defines the parent node
name and each character string a valid attribute name.}

\item{\code{allChildren}}{Character vector, names of globally valid child nodes for all nodes, if any.}

\item{\code{allAttrs}}{Character vector, names of globally valid attributes for all nodes, if any.}

\item{\code{empty}}{Character vector, names of nodes that must be empty nodes (i.e., no closing tag), if any.}
}}
\seealso{
\code{\link[XiMpLe:XMLValidity]{XMLValidity}},
   \code{\link[XiMpLe:validXML]{validXML}}
}
\keyword{classes}

