\docType{methods}
\name{XMLName}
\alias{XMLAttrs}
\alias{XMLAttrs<-}
\alias{XMLAttrs<-,-methods}
\alias{XMLAttrs,-methods}
\alias{XMLAttrs<-,XiMpLe.node-method}
\alias{XMLAttrs,XiMpLe.node-method}
\alias{XMLChildren}
\alias{XMLChildren<-}
\alias{XMLChildren<-,-methods}
\alias{XMLChildren,-methods}
\alias{XMLChildren<-,XiMpLe.doc-method}
\alias{XMLChildren,XiMpLe.doc-method}
\alias{XMLChildren<-,XiMpLe.node-method}
\alias{XMLChildren,XiMpLe.node-method}
\alias{XMLDecl}
\alias{XMLDecl<-}
\alias{XMLDecl<-,-methods}
\alias{XMLDecl,-methods}
\alias{XMLDecl<-,XiMpLe.doc-method}
\alias{XMLDecl,XiMpLe.doc-method}
\alias{XMLDTD}
\alias{XMLDTD<-}
\alias{XMLDTD<-,-methods}
\alias{XMLDTD,-methods}
\alias{XMLDTD<-,XiMpLe.doc-method}
\alias{XMLDTD,XiMpLe.doc-method}
\alias{XMLFile}
\alias{XMLFile<-}
\alias{XMLFile<-,-methods}
\alias{XMLFile,-methods}
\alias{XMLFile<-,XiMpLe.doc-method}
\alias{XMLFile,XiMpLe.doc-method}
\alias{XMLName}
\alias{XMLName<-}
\alias{XMLName<-,-methods}
\alias{XMLName,-methods}
\alias{XMLName<-,XiMpLe.node-method}
\alias{XMLName,XiMpLe.node-method}
\alias{XMLScan}
\alias{XMLScan<-}
\alias{XMLScan<-,-methods}
\alias{XMLScan,-methods}
\alias{XMLScan<-,XiMpLe.doc-method}
\alias{XMLScan,XiMpLe.doc-method}
\alias{XMLScan<-,XiMpLe.node-method}
\alias{XMLScan,XiMpLe.node-method}
\alias{XMLValue}
\alias{XMLValue<-}
\alias{XMLValue<-,-methods}
\alias{XMLValue,-methods}
\alias{XMLValue<-,XiMpLe.node-method}
\alias{XMLValue,XiMpLe.node-method}
\title{Getter/setter methods for S4 objects of XiMpLe XML classes}
\arguments{
  \item{obj}{An object of class \code{XiMpLe.node} or
  \code{XiMpLe.doc}}
}
\description{
  Used to get/set certain slots from objects of class
  \code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}} and
  \code{\link[XiMpLe:XiMpLe.node-class]{XiMpLe.node}}.
}
\details{
  These are convenience methods to get or set slots from
  XML objects without using the \code{@} operator.
}
\seealso{
  \code{\link[XiMpLe:XiMpLe.doc-class]{XiMpLe.doc}}
  \code{\link[XiMpLe:XiMpLe.node-class]{XiMpLe.node}}
}
\keyword{methods}

